makes sure were not delting users who HAVE signed up; job ready to be tested

+++ app/jobs/check_invite_job.rb
@@ -2,7 +2,7 @@ class CheckInviteJob < ApplicationJob
   queue_as :default
 
   def perform(user)
-    if(user.user_invites.order(:created_at).last.created_at >= DateTime.now - 10)
+    if(user.user_invites.order(:created_at).last.created_at >= DateTime.now - 10 && !user.completed_signup?)
       user.destroy
     end
     # Do something later
